checkers/testdata: remove files needed for Go 1.18
