+++ checkers/testdata/badSyncOnceFunc/negative_tests.go
@@ -1,5 +1,3 @@
-//go:build go1.21
-
 package checker_test
 
 import (
