--- checkers/testdata/newDeref/negative_tests118.go
@@ -1,7 +1,0 @@
-//go:build go1.18
-
-package checker_test
-
-func genericNew[T any]() T {
-	return *new(T)
-}
