+++ checkers/testdata/typeDefFirst/negative_tests.go
@@ -9,3 +9,13 @@ func (r *rec) MethodWithRef() {}
 func (r recv) Method2() {}
 
 func JustFunction() {}
+
+type genericNegativeStruct[T any] struct{ _ []T }
+
+func (_ genericNegativeStruct[T]) Method()         {}
+func (_ *genericNegativeStruct[T]) MethodWithRef() {}
+
+type multiGenericNegativeStruct[T any, X any] struct{ _ []T }
+
+func (_ multiGenericNegativeStruct[T, X]) Method()         {}
+func (_ *multiGenericNegativeStruct[T, X]) MethodWithRef() {}
